refactor: selection dispatch, helper, query
