--- src/helper/selection.ts
@@ -1,19 +1,0 @@
-import { SelectionRange, Range } from '../store/types';
-import { isNull } from './common';
-
-export function isSameInputRange(inp1: SelectionRange | null, inp2: SelectionRange | null) {
-  if (isNull(inp1) || isNull(inp2)) {
-    return inp1 === inp2;
-  }
-
-  return (
-    inp1.column[0] === inp2.column[0] &&
-    inp1.column[1] === inp2.column[1] &&
-    inp1.row[0] === inp2.row[0] &&
-    inp1.row[1] === inp2.row[1]
-  );
-}
-
-export function getSortedRange(range: Range): Range {
-  return range[0] > range[1] ? [range[1], range[0]] : range;
-}
